\name{simda}
\alias{simda}
\docType{data}
\title{A simulated data with four individual outcomes}
\description{A simulated data with with four individual outcomes}
\format{
  A data frame with 200 observations on the following 8 variables.
  \describe{
    \item{\code{Y1}}{a continuous individual outcome.}
    \item{\code{Y2}}{a continuous individual outcome.}
    \item{\code{Y3}}{a binary individual outcome.}
    \item{\code{Y4}}{a multinomial individual outcome.}
    \item{\code{Z1}}{a covariate.}
    \item{\code{Z2}}{a covariate.}
    \item{\code{Z3}}{a covariate.}
    \item{\code{Z4}}{a covariate.}
  }
}
\usage{data(simda)}
\keyword{datasets}
